weight int

+++ Sources/GroceryAppShareDTO/ExerciseRequestDTO.swift
@@ -12,9 +12,9 @@ public struct ExerciseRequestDTO: Codable {
     public let gender: String
     
     public let age: Int
-    public let weight: Float
+    public let weight: Int
     
-    public init(gender: String, age:Int,  weight: Float) {
+    public init(gender: String, age:Int,  weight: Int) {
        // self.id = id
         self.gender = gender
